// src/app/Models/provider-office.model.ts
export class ProviderOfficeModel {
    id: number;
    provider_id: number;
    provider: any;
    office: string;
    office_contact: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    phone1: string;
    phone2: string;
    fax: string;
    web_url: string;
}